Phase 2 slice A: core::ISource — StreamSource + CDSource implement the internal Source interface
include/core/ISource.h (new): the compile-time playback-source contract —
readFrames (the audio-callback contract, fixed 44100/stereo/f32) / caps
(seekable, finite, live) / positionSec / durationSec / seekTo / close.
Statically linked, NOT a plugin ABI (Phase 4). open(), metadata, and pause are
deliberately excluded (roadmap Decisions log); CD's extended surface (TOC,
drive offset/model, stopReader, media checks) stays concrete on CDSource.
Header is platform-free (standalone-compile verified, the IHttp seam test).

StreamSource: inherits ISource; override on readFrames/close; positionSec
widened int->double (whole seconds, byte-equivalent); caps=live,
durationSec=0, seekTo=false. src/StreamSource.cpp: ZERO diff — producers,
ring, re-pin, rawRead, and every hls function untouched.

CDSource: inherits ISource; caps=seekable+finite; three signature
harmonizations argued byte-equivalent: seekTo(int)->bool seekTo(double)
(LBA truncation identical for integral input; false on the existing
nothing-playing early-out), positionSec/durationSec int->double (integer
division preserved, widened). readerWorker/ring/readFrames/readSectors: no hunks.

AudioManager: exactly 4 consumer casts (seekBy CD branch, cdPositionSec/
cdDurationSec, streamPositionSec). UIManager + CDRipper compile unchanged —
nothing outside AudioManager called the harmonized signatures.

Tests: cd_pipeline_test + hls_pipeline_test retargeted through core::ISource&
(readChunk/drainFrames dispatch via the interface, so the slice-0 fidelity/
prebuffer assertions now also prove interface dispatch) + new contract blocks
(caps declared-not-faked, seekTo/durationSec per capability).

Audit: sacred-symbol grep over the full diff (producerWorker, readerWorker,
ringWrite/Read/Clear, rawRead, readAudio, InternetReadFile/OpenUrl,
hlsRawRead/Fetch/Poll, prebuffered_, hls_repin) = zero hits. Virtual dispatch
added only on UI-thread query paths — the audio callback still reads through
the concrete members (dispatch there is slice C, deliberately deferred).
Build clean, ctest 13/13, repeated-run stable.

+++ include/AudioManager.h
@@ -142,7 +142,7 @@ public:
     std::string streamNowPlaying() const { return stream_source_.nowPlaying(); }
     std::string streamArtUrl()     const { return stream_source_.currentArtUrl(); } // iHeart digital cover ("" -> use logo)
     std::string streamUrl()      const { return stream_source_.url(); }   // URL actually streaming
-    int      streamPositionSec() const { return stream_source_.positionSec(); }
+    int      streamPositionSec() const { return (int)stream_source_.positionSec(); }
 #endif
 
     // Called by track-end callback to pre-load next track for crossfade/gapless
